+++ travis/script.sh
@@ -1,7 +1,6 @@
 #!/usr/bin/env bash
-
 if [ $PYTHON_INSTALL_METHOD == "tox" ]; then
-  env LDFLAGS="-L$(brew --prefix openssl)/lib" CFLAGS="-I$(brew --prefix openssl)/include" tox
+  env LDFLAGS="-L$(brew --prefix openssl)/lib" CFLAGS="-I$(brew --prefix openssl)/include" tox -e $(tox -l | grep $PYTHON_MAJOR | tr "\n" ",")
 else
   python setup.py install
   coverage run --source werobot -m py.test