+++ werobot/crypto/__init__.py
@@ -10,8 +10,8 @@ try:
     from cryptography.hazmat.primitives.ciphers import \
         Cipher, algorithms, modes
     from cryptography.hazmat.backends import default_backend
-except ImportError:
-    raise RuntimeError("You need to install Cryptography.")
+except ImportError:  # pragma: no cover
+    raise RuntimeError("You need to install Cryptography.")  # pragma: no cover
 
 from . import pkcs7
 from .exceptions import (